Code cleanup
1.  I/O optimization (short of ISR pipelining).  Determined that the best
    scan cycle achievable is approx 3 ms.  With key switch transitions from
    3 ms to 30 ms in steps of 3 ms, this is not sufficient to scan into
    note-on velocities ranging from 1 to 127.
2.  Clean up more compile (rid of warnings).
3.  Minor I/O optimization.

Note: This closes out stage 2 development.  Next stage will be after a change
      to the hardware.  The I2C bus appears to be bottleneck here.

+++ Ad7997.cpp
@@ -44,7 +44,7 @@ uint32_t CAd7997::sync(void)
 
 uint32_t CAd7997::sync(uint8_t i)
 {
-  uint8_t data_wr[] = { E_READ_CMD + (i << 4) };
+  uint8_t data_wr[] = { (uint8_t)(E_READ_CMD + (i << 4)) };
   twi_initiate_transaction(m_i2cAddr, data_wr, sizeof(data_wr), m_inBytes + (2 * i), sizeof(m_inBytes[0]) * 2, &m_readAt);
   twi_wait_until_master_ready();
   return m_readAt;
